Move constants from file into class
